requirements.txt
imageai==3.0.3
inflect==7.0.0
keras_ocr==0.9.2
streamlit==1.25.0
nltk==3.8.1
opencv_python==4.8.0.74
pandas==2.0.3
Pillow==10.0.0
pytorch_lightning==2.0.4
scikit_learn==1.3.0
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
ultralytics==8.0.122
smart_open==6.3.0
gTTS==2.2.3
numpy==1.23.3
python-dotenv==1.0.0
sentencepiece==0.1.99
happytransformer==2.4.1
imageai
inflect
